+++ zellij-server/src/route.rs
@@ -2627,6 +2627,18 @@ pub(crate) fn route_thread_main(
                             )
                             .with_context(err_context)?;
                         },
+                        ClientToServerMsg::KittyZlibSupport { supported } => {
+                            send_to_screen_or_retry_queue!(
+                                senders,
+                                ScreenInstruction::SetKittyZlibSupport {
+                                    client_id,
+                                    supported
+                                },
+                                instruction,
+                                retry_queue
+                            )
+                            .with_context(err_context)?;
+                        },
                         ClientToServerMsg::SixelSupport { supported } => {
                             send_to_screen_or_retry_queue!(
                                 senders,