+++ zellij-utils/src/ipc/tests/roundtrip_tests.rs
@@ -362,6 +362,8 @@ fn test_client_messages() {
     });
     test_client_roundtrip!(ClientToServerMsg::KittyGraphicsSupport { supported: true });
     test_client_roundtrip!(ClientToServerMsg::KittyGraphicsSupport { supported: false });
+    test_client_roundtrip!(ClientToServerMsg::KittyZlibSupport { supported: true });
+    test_client_roundtrip!(ClientToServerMsg::KittyZlibSupport { supported: false });
     test_client_roundtrip!(ClientToServerMsg::SixelSupport { supported: true });
     test_client_roundtrip!(ClientToServerMsg::SixelSupport { supported: false });
     test_client_roundtrip!(ClientToServerMsg::BackgroundColor {